Refactor the select call into two separate calls
The first call is to prepare the query. At the moment, this runs some
basic modifications to the AST. In the future, I want it to also run
validations so that those validations get moved out of the parser.

The second is the call to select which operates the same as it used to
operate.